pyproject.toml
[tool.poetry]
name = "bitcoin_in_python"
version = "0.1.0"
description = ""
authors = ["cjc7373 <niuchangcun@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tinydb = "^4.5.2"
pycryptodome = "^3.11.0"
base58 = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
